Set up Vitest configuration and testing utilities with mocks for streamlining translation cleanup tests.
