+++ package.json
@@ -2,6 +2,7 @@
   "name": "vue-translations-cleanup",
   "version": "1.0.3",
   "description": "A tool to find and remove unused translation keys in Vue i18n projects",
+  "type": "module",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "bin": {
