+++ tests/translations-cleanup/index.test.ts
@@ -1,51 +1,9 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
-import { cleanupTranslations } from '../src/translation-cleanup'
+import {describe, expect, it, vi} from 'vitest'
+import {cleanupTranslations} from '../../src/translations-cleanup'
 import * as fs from 'node:fs'
-import { glob } from 'glob'
-
-// Mock fs and glob
-vi.mock('node:fs')
-vi.mock('glob')
+import {mockTranslations} from '../setup'
 
 describe('cleanupTranslations', () => {
-    const mockTranslations = {
-        common: {
-            hello: 'Hello',
-            unused: 'Unused Key'
-        },
-        buttons: {
-            submit: 'Submit'
-        }
-    }
-
-    const mockFiles = [
-        'src/components/Test.vue',
-        'src/utils/helper.ts'
-    ]
-
-    beforeEach(() => {
-        // Setup fs mock
-        vi.mocked(fs.existsSync).mockImplementation(() => true)
-        vi.mocked(fs.readFileSync).mockImplementation((path) => {
-            if (path === 'translations.json') {
-                return JSON.stringify(mockTranslations)
-            }
-            // Mock file content that uses some translations
-            return `
-                const t = (key) => key;
-                t('common.hello')
-                t('buttons.submit')
-            `
-        })
-
-        // Setup glob mock
-        vi.mocked(glob).mockResolvedValue(mockFiles)
-    })
-
-    afterEach(() => {
-        vi.clearAllMocks()
-    })
-
     it('should validate input paths', async () => {
         vi.mocked(fs.existsSync).mockImplementation(() => false)
