+++ vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+    test: {
+        include: ['tests/**/*.test.{ts,js}'],
+        setupFiles: ['./tests/setup.ts'],
+    },
+})
